requirements.txt
python-dotenv==0.19.0
requests==2.26.0
python-coinmarketcap==0.3
